Rename SyncGitHubReleases to SyncReleases
