+++ src/Squirrel.nuspec
@@ -26,7 +26,7 @@
     <file src="Setup\bin\Release\Setup.exe" target="tools" />
     <file src="Update\bin\Release\Update.exe" target="tools\Squirrel.exe" />
     <file src="Update\bin\Release\Update.com" target="tools\Squirrel.com" />
-    <file src="SyncGitHubReleases\bin\Release\SyncGitHubReleases.exe" target="tools" />
+    <file src="SyncReleases\bin\Release\SyncReleases.exe" target="tools" />
     <file src="Update\signtool.exe" target="tools\signtool.exe" />
     <file src="Update\rcedit.exe" target="tools\rcedit.exe" />
   </files>
