+++ src/SyncReleases/App.config
