+++ src/SyncReleases/Properties/AssemblyInfo.cs
@@ -3,9 +3,9 @@ using System.Runtime.InteropServices;
 
 // All other assembly info is defined in SharedAssembly.cs
 
-[assembly: AssemblyTitle("SyncGitHubReleases")]
-[assembly: AssemblyProduct("SyncGitHubReleases")]
-[assembly: AssemblyDescription("SyncGitHubReleases")]
+[assembly: AssemblyTitle("SyncReleases")]
+[assembly: AssemblyProduct("SyncReleases")]
+[assembly: AssemblyDescription("SyncReleases")]
 
 // Setting ComVisible to false makes the types in this assembly not visible 
 // to COM components.  If you need to access a type in this assembly from
