+++ src/SyncReleases/packages.config
